ADD: frame - moving up in frames
added some logic to the roll method, with a new score method, the
score is now an array of arrays so we need to flatten that array first
before reducing the arrays to get a score so have refacotred that. also
changed the init so have a couple of helpers to iterate the roll and
frame in the array of arrays.